Add first database table page
The first database table is implemented for the first link of the menu-bar. It is also accessible by going straight to localhost:3000/table. The table is dynamically drawn by using template loops to go through the content. However, we need to make this system more dynamic by querying without knowing all the attributes prior. Not doing so would cause a lot of breaches to DRY principles. However, I will commit for now for demonstration purposes.