Add dynamic routing for tables and edited menu-bar link
The tables can now be routed using /tables/*. From the menu-bar, we can add links to specific tables by specifying the another routing parameter. i.e. /tables/prodhistory/ will link to the Product History. On failure to find a table, a 404 will be thrown. However, be warned that this produced another level to the hierarchy (the reason all resource links have ‘..’ added to then to specify another layer down the hierarchy)

I also edited the title to be a link to the root page in the menu bar.