Add front-end implementation of dynamic table editing
The front-end of the CRUD features have been implemented into the table pages. They are dynamically generated using the HTML that exists on the page — Don’t see much of a problem with this since we generate these tables in a pretty predictable pattern. However, doing this such as adding custom columns that exists not for data display (ex. Original implementation of the CRUD interface) will not work as it will offset the data matching.